updated binned SED generation functionality
